Begin splitting command_queue_interface.hpp into header and implementation
