+++ tt_metal/api/tt-metalium/command_queue_interface.hpp
@@ -255,78 +255,16 @@ inline uint32_t get_absolute_cq_offset(uint16_t channel, uint8_t cq_id, uint32_t
 }
 
 template <bool addr_16B>
-inline uint32_t get_cq_issue_rd_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size) {
-    uint32_t recv;
-    chip_id_t mmio_device_id = tt::Cluster::instance().get_associated_mmio_device(chip_id);
-    uint16_t channel = tt::Cluster::instance().get_assigned_channel_for_device(chip_id);
-    uint32_t channel_offset = (channel >> 2) * tt::tt_metal::DispatchSettings::MAX_DEV_CHANNEL_SIZE;
-    CoreType core_type = tt::tt_metal::dispatch_core_manager::instance().get_dispatch_core_type(chip_id);
-    uint32_t issue_q_rd_ptr =
-        DispatchMemMap::get(core_type).get_host_command_queue_addr(CommandQueueHostAddrType::ISSUE_Q_RD);
-    tt::Cluster::instance().read_sysmem(
-        &recv,
-        sizeof(uint32_t),
-        issue_q_rd_ptr + channel_offset + get_relative_cq_offset(cq_id, cq_size),
-        mmio_device_id,
-        channel);
-    if (not addr_16B) {
-        return recv << 4;
-    }
-    return recv;
-}
+uint32_t get_cq_issue_rd_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size);
 
 template <bool addr_16B>
-inline uint32_t get_cq_issue_wr_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size) {
-    uint32_t recv;
-    chip_id_t mmio_device_id = tt::Cluster::instance().get_associated_mmio_device(chip_id);
-    uint16_t channel = tt::Cluster::instance().get_assigned_channel_for_device(chip_id);
-    CoreType core_type = tt::tt_metal::dispatch_core_manager::instance().get_dispatch_core_type(chip_id);
-    uint32_t issue_q_wr_ptr =
-        DispatchMemMap::get(core_type).get_host_command_queue_addr(CommandQueueHostAddrType::ISSUE_Q_WR);
-    tt::Cluster::instance().read_sysmem(
-        &recv, sizeof(uint32_t), issue_q_wr_ptr + get_relative_cq_offset(cq_id, cq_size), mmio_device_id, channel);
-    if (not addr_16B) {
-        return recv << 4;
-    }
-    return recv;
-}
+uint32_t get_cq_issue_wr_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size);
 
 template <bool addr_16B>
-inline uint32_t get_cq_completion_wr_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size) {
-    uint32_t recv;
-    chip_id_t mmio_device_id = tt::Cluster::instance().get_associated_mmio_device(chip_id);
-    uint16_t channel = tt::Cluster::instance().get_assigned_channel_for_device(chip_id);
-    uint32_t channel_offset = (channel >> 2) * tt::tt_metal::DispatchSettings::MAX_DEV_CHANNEL_SIZE;
-    CoreType core_type = tt::tt_metal::dispatch_core_manager::instance().get_dispatch_core_type(chip_id);
-    uint32_t completion_q_wr_ptr =
-        DispatchMemMap::get(core_type).get_host_command_queue_addr(CommandQueueHostAddrType::COMPLETION_Q_WR);
-    tt::Cluster::instance().read_sysmem(
-        &recv,
-        sizeof(uint32_t),
-        completion_q_wr_ptr + channel_offset + get_relative_cq_offset(cq_id, cq_size),
-        mmio_device_id,
-        channel);
-    if (not addr_16B) {
-        return recv << 4;
-    }
-    return recv;
-}
+uint32_t get_cq_completion_wr_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size);
 
 template <bool addr_16B>
-inline uint32_t get_cq_completion_rd_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size) {
-    uint32_t recv;
-    chip_id_t mmio_device_id = tt::Cluster::instance().get_associated_mmio_device(chip_id);
-    uint16_t channel = tt::Cluster::instance().get_assigned_channel_for_device(chip_id);
-    CoreType core_type = tt::tt_metal::dispatch_core_manager::instance().get_dispatch_core_type(chip_id);
-    uint32_t completion_q_rd_ptr =
-        DispatchMemMap::get(core_type).get_host_command_queue_addr(CommandQueueHostAddrType::COMPLETION_Q_RD);
-    tt::Cluster::instance().read_sysmem(
-        &recv, sizeof(uint32_t), completion_q_rd_ptr + get_relative_cq_offset(cq_id, cq_size), mmio_device_id, channel);
-    if (not addr_16B) {
-        return recv << 4;
-    }
-    return recv;
-}
+uint32_t get_cq_completion_rd_ptr(chip_id_t chip_id, uint8_t cq_id, uint32_t cq_size);
 
 struct SystemMemoryCQInterface {
     // CQ is split into issue and completion regions
@@ -412,94 +350,7 @@ private:
         worker_launch_message_buffer_state;
 
 public:
-    SystemMemoryManager(chip_id_t device_id, uint8_t num_hw_cqs) :
-        device_id(device_id),
-        num_hw_cqs(num_hw_cqs),
-        fast_write_callable(tt::Cluster::instance().get_fast_pcie_static_tlb_write_callable(device_id)),
-        bypass_enable(false),
-        bypass_buffer_write_offset(0) {
-        this->completion_byte_addrs.resize(num_hw_cqs);
-        this->prefetcher_cores.resize(num_hw_cqs);
-        this->prefetch_q_writers.reserve(num_hw_cqs);
-        this->prefetch_q_dev_ptrs.resize(num_hw_cqs);
-        this->prefetch_q_dev_fences.resize(num_hw_cqs);
-
-        // Split hugepage into however many pieces as there are CQs
-        chip_id_t mmio_device_id = tt::Cluster::instance().get_associated_mmio_device(device_id);
-        uint16_t channel = tt::Cluster::instance().get_assigned_channel_for_device(device_id);
-        char* hugepage_start = (char*)tt::Cluster::instance().host_dma_address(0, mmio_device_id, channel);
-        hugepage_start += (channel >> 2) * DispatchSettings::MAX_DEV_CHANNEL_SIZE;
-        this->cq_sysmem_start = hugepage_start;
-
-        // TODO(abhullar): Remove env var and expose sizing at the API level
-        char* cq_size_override_env = std::getenv("TT_METAL_CQ_SIZE_OVERRIDE");
-        if (cq_size_override_env != nullptr) {
-            uint32_t cq_size_override = std::stoi(string(cq_size_override_env));
-            this->cq_size = cq_size_override;
-        } else {
-            this->cq_size = tt::Cluster::instance().get_host_channel_size(mmio_device_id, channel) / num_hw_cqs;
-            if (tt::Cluster::instance().is_galaxy_cluster()) {
-                // We put 4 galaxy devices per huge page since number of hugepages available is less than number of
-                // devices.
-                this->cq_size = this->cq_size / DispatchSettings::DEVICES_PER_UMD_CHANNEL;
-            }
-        }
-        this->channel_offset = DispatchSettings::MAX_HUGEPAGE_SIZE * get_umd_channel(channel) + (channel >> 2) * DispatchSettings::MAX_DEV_CHANNEL_SIZE;
-
-        CoreType core_type = tt::tt_metal::dispatch_core_manager::instance().get_dispatch_core_type(device_id);
-        uint32_t completion_q_rd_ptr =
-            DispatchMemMap::get(core_type).get_device_command_queue_addr(CommandQueueDeviceAddrType::COMPLETION_Q_RD);
-        uint32_t prefetch_q_base =
-            DispatchMemMap::get(core_type).get_device_command_queue_addr(CommandQueueDeviceAddrType::UNRESERVED);
-        uint32_t cq_start =
-            DispatchMemMap::get(core_type).get_host_command_queue_addr(CommandQueueHostAddrType::UNRESERVED);
-        for (uint8_t cq_id = 0; cq_id < num_hw_cqs; cq_id++) {
-            tt_cxy_pair prefetcher_core =
-                tt::tt_metal::dispatch_core_manager::instance().prefetcher_core(device_id, channel, cq_id);
-            auto prefetcher_virtual = tt::Cluster::instance().get_virtual_coordinate_from_logical_coordinates(prefetcher_core.chip, CoreCoord(prefetcher_core.x, prefetcher_core.y), core_type);
-            this->prefetcher_cores[cq_id] = tt_cxy_pair(prefetcher_core.chip, prefetcher_virtual.x, prefetcher_virtual.y);
-            this->prefetch_q_writers.emplace_back(
-                tt::Cluster::instance().get_static_tlb_writer(this->prefetcher_cores[cq_id]));
-
-            tt_cxy_pair completion_queue_writer_core =
-                tt::tt_metal::dispatch_core_manager::instance().completion_queue_writer_core(device_id, channel, cq_id);
-            auto completion_queue_writer_virtual =
-                tt::Cluster::instance().get_virtual_coordinate_from_logical_coordinates(
-                    completion_queue_writer_core.chip,
-                    CoreCoord(completion_queue_writer_core.x, completion_queue_writer_core.y),
-                    core_type);
-
-            const std::tuple<uint32_t, uint32_t> completion_interface_tlb_data =
-                tt::Cluster::instance()
-                    .get_tlb_data(tt_cxy_pair(
-                        completion_queue_writer_core.chip,
-                        completion_queue_writer_virtual.x,
-                        completion_queue_writer_virtual.y))
-                    .value();
-            auto [completion_tlb_offset, completion_tlb_size] = completion_interface_tlb_data;
-            this->completion_byte_addrs[cq_id] = completion_tlb_offset + completion_q_rd_ptr % completion_tlb_size;
-
-            this->cq_interfaces.push_back(SystemMemoryCQInterface(channel, cq_id, this->cq_size, cq_start));
-            // Prefetch queue acts as the sync mechanism to ensure that issue queue has space to write, so issue queue
-            // must be as large as the max amount of space the prefetch queue can specify Plus 1 to handle wrapping Plus
-            // 1 to allow us to start writing to issue queue before we reserve space in the prefetch queue
-            TT_FATAL(
-                DispatchMemMap::get(core_type, num_hw_cqs).max_prefetch_command_size() *
-                        (DispatchMemMap::get(core_type, num_hw_cqs).prefetch_q_entries() + 2) <=
-                    this->get_issue_queue_size(cq_id),
-                "Issue queue for cq_id {} has size of {} which is too small",
-                cq_id,
-                this->get_issue_queue_size(cq_id));
-            this->cq_to_event.push_back(0);
-            this->cq_to_last_completed_event.push_back(0);
-            this->prefetch_q_dev_ptrs[cq_id] = prefetch_q_base;
-            this->prefetch_q_dev_fences[cq_id] =
-                prefetch_q_base + DispatchMemMap::get(core_type, num_hw_cqs).prefetch_q_entries() *
-                                      sizeof(DispatchSettings::prefetch_q_entry_type);
-        }
-        std::vector<std::mutex> temp_mutexes(num_hw_cqs);
-        cq_to_event_locks.swap(temp_mutexes);
-    }
+    SystemMemoryManager(chip_id_t device_id, uint8_t num_hw_cqs);
 
     uint32_t get_next_event(const uint8_t cq_id) {
         cq_to_event_locks[cq_id].lock();
